Avoid cloudflare captcha

+++ update.sh
@@ -25,28 +25,38 @@ else
     exit 1
 fi
 
-# Generate the new config file.
-echo "# CloudFlare IP Ranges" > $CF_NGINX_CONFIG
-echo "# Generated at $(date) by $0" >> $CF_NGINX_CONFIG
 
-echo "" >> $CF_NGINX_CONFIG
+# Check number of lines. Sometimes CloudFlare shows captcha
+CF_TEMP_IP4_LINES=$(wc -l < $CF_TEMP_IP4)
+CF_TEMP_IP6_LINES=$(wc -l < $CF_TEMP_IP6)
 
-echo "# - IPv4 ($CF_URL_IP4)" >> $CF_NGINX_CONFIG
-awk '{ print "set_real_ip_from " $0 ";" }' $CF_TEMP_IP4 >> $CF_NGINX_CONFIG
+MAX_LINES=20
+if [ $CF_TEMP_IP4_LINES -le $MAX_LINES ] && [ $CF_TEMP_IP6_LINES --le $MAX_LINES ]; then
 
-echo "" >> $CF_NGINX_CONFIG
+  # Generate the new config file.
+  echo "# CloudFlare IP Ranges" > $CF_NGINX_CONFIG
+  echo "# Generated at $(date) by $0" >> $CF_NGINX_CONFIG
 
-echo "# - IPv6 ($CF_URL_IP6)" >> $CF_NGINX_CONFIG
-awk '{ print "set_real_ip_from " $0 ";" }' $CF_TEMP_IP6 >> $CF_NGINX_CONFIG
+  echo "" >> $CF_NGINX_CONFIG
 
-echo "" >> $CF_NGINX_CONFIG
+  echo "# - IPv4 ($CF_URL_IP4)" >> $CF_NGINX_CONFIG
+  awk '{ print "set_real_ip_from " $0 ";" }' $CF_TEMP_IP4 >> $CF_NGINX_CONFIG
 
-echo "real_ip_header CF-Connecting-IP;" >> $CF_NGINX_CONFIG
+  echo "" >> $CF_NGINX_CONFIG
 
-echo "" >> $CF_NGINX_CONFIG
+  echo "# - IPv6 ($CF_URL_IP6)" >> $CF_NGINX_CONFIG
+  awk '{ print "set_real_ip_from " $0 ";" }' $CF_TEMP_IP6 >> $CF_NGINX_CONFIG
 
-# Remove the temporary files.
-rm $CF_TEMP_IP4 $CF_TEMP_IP6
+  echo "" >> $CF_NGINX_CONFIG
 
-# Reload the nginx config.
-/usr/sbin/nginx -t && /usr/sbin/service nginx reload
+  echo "real_ip_header CF-Connecting-IP;" >> $CF_NGINX_CONFIG
+
+  echo "" >> $CF_NGINX_CONFIG
+
+  # Remove the temporary files.
+  rm $CF_TEMP_IP4 $CF_TEMP_IP6
+
+  # Reload the nginx config.
+  /usr/sbin/nginx -t && /usr/sbin/service nginx reload || rm $CF_NGINX_CONFIG
+
+fi
